0.0.3

+++ DevBackEnd.Entities/Concrete/City.cs
@@ -4,8 +4,8 @@ namespace DevBackEnd.Entities.Concrete
 {
     public class City:IEntity
     {
-        public virtual int CityId { get; set; }
-        public virtual byte CountryId { get; set; }
+        public virtual short CityId { get; set; }
+        public virtual int CountryId { get; set; }
         public virtual string CityName { get; set; }
     }
 }
