+++ DevBackEnd.Entities/Concrete/Country.cs
@@ -4,7 +4,7 @@ namespace DevBackEnd.Entities.Concrete
 {
     public class Country : IEntity
     {
-        public virtual int CountryId { get; set; }
+        public virtual byte CountryId { get; set; }
         public virtual string CountryName { get; set; }
     }
 }
